refactor(filter): move Filter files into folder
