+++ packages/components/src/Filter/index.ts
@@ -1,1 +1,2 @@
 export * from "./Filter"
+export * from "./FilterButton"
